pyproject.toml
[project]
name = "fake-geotag"
version = "0.1.0"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "rich",
    "pillow",
    "piexif"
]
